+++ src/test/issueTests.ts
@@ -370,4 +370,12 @@ export const issueTests: {
             ["typedoc"]
         );
     },
+
+    gh1907(_project, logger) {
+        logger.expectMessage(
+            'warn: The --name option was not specified, and package.json does not have a name field. Defaulting project name to "Documentation"'
+        );
+        logger.discardDebugMessages();
+        logger.expectNoOtherMessages();
+    },
 };